Fix for BT issue in Tasmania hw
In the specific hardware configuration power is cut to
the Bluetooth controller every time system enter S3 suspend

Android uses hci0 interface to initialise bluetooth by default.
In some cases when frequent usb connect and disconnect
occurs due to suspend and resume, there is a chance that hci
dev registration happens even before cleanup is done for the
existing hci device, which causes creation of hci1 device.
Changes added to allow userspace to use hci1 when hci0 is not
available

Tracked-On: OAM-99203
Signed-off-by: Aiswarya Cyriac aiswarya.cyriac@intel.com